calc filled check in bitsize_internal and move it to Bitsized trait
