+++ bilge-impl/src/bitsize_internal/struct_gen.rs
@@ -184,7 +184,7 @@ pub(crate) fn generate_getter_inner(ty: &Type, is_getter: bool) -> TokenStream {
                     quote! { {
                         #elem_value
                         // so, has try_from impl
-                        if !#ty::FILLED {
+                        if !<#ty as Bitsized>::FILLED {
                             #ty::try_from(elem_value).is_ok()
                         } else {
                             true
